Added Web Page project
This is a school project made in group together with Saara Lehtimäki and Katja Ikonen. We designed a guest book for rental cottage. 
This project works around databases and PHP. Used also MySQL to create and store DataBase. We used also htAccess for safer Admin Access into web page. 
My part of the project:
- index.php
- LueTietokannasta.php
- Virheilmoitus.html